# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gerenciamento de clientes, agendamentos, orçamentos e ordens de serviço.",
        "title": "API VentuFrio",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte VentuFrio",
            "email": "suporte@ventufrio.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/ordens": {
            "get": {
                "description": "Retorna todas as ordens de serviço cadastradas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordens-de-servico"
                ],
                "summary": "Lista todas as ordens de serviço",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.OrdemDeServico"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Registra uma nova ordem de serviço com os dados fornecidos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordens-de-servico"
                ],
                "summary": "Cria uma nova ordem de serviço",
                "parameters": [
                    {
                        "description": "Dados da Ordem de Serviço",
                        "name": "ordem",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.OrdemDeServico"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/ordens/{id}": {
            "get": {
                "description": "Retorna os dados da ordem de serviço com base no ID fornecido",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordens-de-servico"
                ],
                "summary": "Busca uma ordem de serviço pelo ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da Ordem de Serviço",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.OrdemDeServico"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Atualiza os dados de uma ordem de serviço com base no ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordens-de-servico"
                ],
                "summary": "Atualiza uma ordem de serviço existente",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da Ordem de Serviço",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Novos dados da Ordem de Serviço",
                        "name": "ordem",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.OrdemDeServico"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove uma ordem de serviço com base no ID fornecido",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordens-de-servico"
                ],
                "summary": "Exclui uma ordem de serviço",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da Ordem de Serviço",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Agendamento": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "data_visita": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "orcamentos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Orcamento"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Cliente": {
            "type": "object",
            "properties": {
                "bairro": {
                    "type": "string"
                },
                "cep": {
                    "type": "string"
                },
                "cpf": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "log_delete": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "numero": {
                    "type": "string"
                },
                "orcamentos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Orcamento"
                    }
                },
                "rua": {
                    "type": "string"
                },
                "sobrenome": {
                    "type": "string"
                },
                "telefone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Material": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "ordem_de_servico": {
                    "$ref": "#/definitions/models.OrdemDeServico"
                },
                "ordem_de_servico_id": {
                    "type": "integer"
                },
                "quantidade": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Orcamento": {
            "type": "object",
            "properties": {
                "agendamento": {
                    "$ref": "#/definitions/models.Agendamento"
                },
                "agendamento_id": {
                    "type": "integer"
                },
                "cliente": {
                    "$ref": "#/definitions/models.Cliente"
                },
                "cliente_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "delete_log": {
                    "type": "string"
                },
                "descricao_inicial": {
                    "type": "string"
                },
                "descricao_item": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ordens_de_servico": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.OrdemDeServico"
                    }
                },
                "prazo_entrega": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.OrdemDeServico": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "descricao_servico": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "materiais": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Material"
                    }
                },
                "orcamento": {
                    "$ref": "#/definitions/models.Orcamento"
                },
                "orcamento_id": {
                    "type": "integer"
                },
                "responsavel": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}